Use symbolLevel=none only for English; for other languages, inherit it from the level below. (#4)
Fixes nvaccess/nvda#14417

Summary of the issue:
Hindi has no symbol defined in its symbol file, only copyright header; seems that the file was prepared for translation but no actual symbol translation took place.
But there is a Hindi CLDR file. Thus the symbol level for symbols such as common punctuation (dot, question marke, etc.) is the one of CLDR, i.e. none.

This is not adapted and it would be better to take advantage of the symbol levels that are defined in the English symbol file.

Description of user facing changes
In NVDA, locale CLDR dic file inherits symbol levels from the files coming after, i.e. English symbols and English CLDR. In case the locale symbol file does not define a character's level, this allows to:

Use the level for this symbol if it is defined there
Use "none" (coming from English CLDR dic file) if the character is not defined in English symbol file but is defined in CLDR.
Description of development approach
For all languages except English ("en"), generate the cldr.dic file with "-" in the level field, meaning that the level is inherited from previous files.
For English cldr.dic file, use "none" for the symbol level, as it was already before this PR.